requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
Deprecated==1.2.18
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.16.0
idna==3.10
pillow==11.2.1
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyGithub==2.6.1
PyJWT==2.10.1
PyMySQL==1.1.1
PyNaCl==1.5.0
python-dotenv==1.1.0
python-magic==0.4.27
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wrapt==1.17.2
